rename: postData 관련 hook를 hooks/post 폴더로 이동 - api 응답을 Post 컴포넌트에 필요한 형태로 가공하는 인터페이스 재사용을 위해 분리 [feat/21]
