+++ src/hooks/index.js
@@ -7,6 +7,6 @@ export { default as useSocialLogin } from "./useSocialLogin";
 export { default as useWritePost } from "./useWritePost";
 export { default as useTagSelector } from "./useTagSelector";
 export { default as usePreventScroll } from "./usePreventScroll";
-export { default as useProcessPostsData } from "./useProcessPostsData";
 export { default as useCardFooterButtonClick } from "./useCardFooterButtonClick";
 export { default as useInfiniteScroll } from "./useInfiniteScroll";
+export * from "./post";
