+++ src/hooks/post/index.js
@@ -1,0 +1,1 @@
+export { default as useProcessPostsData } from "./useProcessPostsData";
